backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.11.0
python-dotenv>=1.0.0
google-genai>=1.40.0
python-multipart>=0.0.12
